fix(cli): extract quota settlement rollout helpers to restore module-size budget (#3021)
